admin-panel/requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
coverage==7.3.2
crispy-bootstrap5==2023.10
cssbeautifier==1.14.11
Django==5.0
django-attachments==1.11
django-contrib-comments==2.2.0
django-crispy-forms==2.1
django-filter==23.5
django-guardian==2.4.0
django-login-required-middleware==0.9.0
django-stubs==4.2.6
django-stubs-ext==4.2.5
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
djlint==1.34.0
EditorConfig==0.12.3
execnet==2.0.2
fontawesomefree==6.5.0
html-tag-names==0.1.2
html-void-elements==0.1.0
iniconfig==2.0.0
jsbeautifier==1.14.11
json5==0.9.14
mock==5.1.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
pluggy==1.3.0
psutil==5.9.6
pytest==7.4.3
pytest-django==4.7.0
pytest-xdist==3.5.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzdata==2023.3
